requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
build==1.2.1
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
distlib==0.3.8
distro==1.9.0
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.15.3
google==3.0.0
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.8
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
jiter==0.5.0
keyring==24.3.1
more-itertools==10.3.0
msgpack==1.0.8
mysql-connector-python==9.0.0
numpy==2.1.1
openai==1.44.1
packaging==24.1
pandas==2.2.2
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
proto-plus==1.24.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.1
pydantic_core==2.23.3
pyparsing==3.1.4
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.2
rapidfuzz==3.9.3
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
SecretStorage==3.3.3
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tomlkit==0.12.5
tqdm==4.66.5
trove-classifiers==2024.5.22
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
wheel==0.43.0
